Refactor build scripts to allow building only Classical or Mono
The in-container build scripts now get passed CLASSICAL and MONO env
variables which can be used to build one or the other, or both
(default).

`build.sh`, `build-release.sh` and `build-templates.sh` now all expect
command line switches to specify the version details, and optionally
which flavor to build.

For example to build Mono only:

  ./build.sh -v 3.2-beta4 -g master -b mono
  ./build-release.sh -v 3.2-beta4 -b mono
  ./build-templates.sh -v 3.2-beta4 -t 3.2.beta4 -b mono

Also took the opportunity to do some extra cleanup, like removing
unnecessary `builtin_*` options since they all default to True, even
for Linux in 3.2, as well as `use_lto` and `use_static_cpp` options
on platforms which don't implement them.

And I improved the `build-release.sh` script to be a bit easier to
read, and avoid having too many stray folders to cleanup.
The build scripts should now generate the final structure that we'd use
on the official mirrors, with the `mono` distribution as a subfolder of
the main release folder.

+++ build-javascript/build.sh
@@ -2,35 +2,49 @@
 
 set -e
 
+# Config
+
 export BUILD_NAME=official
 export SCONS="scons -j${NUM_CORES} verbose=yes warnings=no progress=no"
-export OPTIONS="builtin_libpng=yes builtin_openssl=yes builtin_zlib=yes debug_symbols=no use_static_cpp=yes use_lto=yes"
+export OPTIONS="debug_symbols=no"
 export OPTIONS_MONO="module_mono_enabled=yes mono_static=yes mono_prefix=/root/mono-installs/wasm-runtime-release"
 export TERM=xterm
-
-source /root/emsdk/emsdk_env.sh
 
 rm -rf godot
 mkdir godot
 cd godot
 tar xf /root/godot.tar.gz --strip-components=1
 
-cp /root/mono-glue/*.cpp modules/mono/glue/
-cp -r /root/mono-glue/Managed/Generated modules/mono/glue/Managed/
+# Classical
 
-$SCONS platform=javascript ${OPTIONS} target=release_debug tools=no
-$SCONS platform=javascript ${OPTIONS} target=release tools=no
+if [ "${CLASSICAL}" == "1" ]; then
+  echo "Starting classical build for JavaScript..."
 
-mkdir -p /root/out/templates
-cp -rvp bin/*.zip /root/out/templates
-rm -f bin/*.zip
+  $SCONS platform=javascript ${OPTIONS} target=release_debug tools=no
+  $SCONS platform=javascript ${OPTIONS} target=release tools=no
 
-$SCONS platform=javascript ${OPTIONS} ${OPTIONS_MONO} target=release_debug tools=no
-$SCONS platform=javascript ${OPTIONS} ${OPTIONS_MONO} target=release tools=no
+  mkdir -p /root/out/templates
+  cp -rvp bin/*.zip /root/out/templates
+  rm -f bin/*.zip
+fi
 
-mkdir -p /root/out/templates-mono
-cp -rvp bin/*.zip /root/out/templates-mono
-rm -f bin/*.zip
+# Mono
 
-mkdir -p /root/out/templates-mono/bcl
-cp -r /root/mono-installs/wasm-bcl/wasm /root/out/templates-mono/bcl/
+if [ "${MONO}" == "1" ]; then
+  echo "Starting Mono build for JavaScript..."
+
+  cp /root/mono-glue/*.cpp modules/mono/glue/
+  cp -r /root/mono-glue/Managed/Generated modules/mono/glue/Managed/
+
+  $SCONS platform=javascript ${OPTIONS} ${OPTIONS_MONO} target=release_debug tools=no
+  $SCONS platform=javascript ${OPTIONS} ${OPTIONS_MONO} target=release tools=no
+
+  mkdir -p /root/out/templates-mono
+  cp -rvp bin/*.zip /root/out/templates-mono
+  rm -f bin/*.zip
+
+  mkdir -p /root/out/templates-mono/bcl
+  cp -r /root/mono-installs/wasm-bcl/wasm /root/out/templates-mono/bcl/
+fi
+
+echo "JavaScript build successful"
