+++ build-macosx/build.sh
@@ -2,9 +2,11 @@
 
 set -e
 
+# Config
+
 export BUILD_NAME=official
 export SCONS="scons -j${NUM_CORES} verbose=yes warnings=no progress=no"
-export OPTIONS="osxcross_sdk=darwin18 builtin_libpng=yes builtin_openssl=yes builtin_zlib=yes debug_symbols=no use_static_cpp=yes use_lto=yes"
+export OPTIONS="osxcross_sdk=darwin18 debug_symbols=no"
 export OPTIONS_MONO="module_mono_enabled=yes mono_static=yes mono_prefix=/root/dependencies/mono"
 export TERM=xterm
 
@@ -13,35 +15,47 @@ mkdir godot
 cd godot
 tar xf /root/godot.tar.gz --strip-components=1
 
-cp /root/mono-glue/*.cpp modules/mono/glue/
-cp -r /root/mono-glue/Managed/Generated modules/mono/glue/Managed/
+# Classical
 
-$SCONS platform=osx $OPTIONS tools=yes target=release_debug
+if [ "${CLASSICAL}" == "1" ]; then
+  echo "Starting classical build for macOS..."
 
-mkdir -p /root/out/tools
-cp -rvp bin/* /root/out/tools
-rm -rf bin
+  $SCONS platform=osx $OPTIONS tools=yes target=release_debug
 
-$SCONS platform=osx $OPTIONS tools=no target=release_debug
-$SCONS platform=osx $OPTIONS tools=no target=release
+  mkdir -p /root/out/tools
+  cp -rvp bin/* /root/out/tools
+  rm -rf bin
 
-mkdir -p /root/out/templates
-cp -rvp bin/* /root/out/templates
-rm -rf bin
+  $SCONS platform=osx $OPTIONS tools=no target=release_debug
+  $SCONS platform=osx $OPTIONS tools=no target=release
 
-$SCONS platform=osx $OPTIONS $OPTIONS_MONO tools=yes target=release_debug copy_mono_root=yes
+  mkdir -p /root/out/templates
+  cp -rvp bin/* /root/out/templates
+  rm -rf bin
+fi
 
-mkdir -p /root/out/tools-mono
-cp -rvp bin/* /root/out/tools-mono
-rm -rf bin
+# Mono
 
-$SCONS platform=osx $OPTIONS $OPTIONS_MONO tools=no target=release_debug
-$SCONS platform=osx $OPTIONS $OPTIONS_MONO tools=no target=release
+if [ "${MONO}" == "1" ]; then
+  echo "Starting Mono build for macOS..."
 
-mkdir -p /root/out/templates-mono
-cp -rvp bin/* /root/out/templates-mono
-rm -rf bin
+  cp /root/mono-glue/*.cpp modules/mono/glue/
+  cp -r /root/mono-glue/Managed/Generated modules/mono/glue/Managed/
 
-find /root/out -name config -exec cp /root/dependencies/mono/etc/config {} \;
+  $SCONS platform=osx $OPTIONS $OPTIONS_MONO tools=yes target=release_debug copy_mono_root=yes
 
-echo "MacOS build successful"
+  mkdir -p /root/out/tools-mono
+  cp -rvp bin/* /root/out/tools-mono
+  rm -rf bin
+
+  $SCONS platform=osx $OPTIONS $OPTIONS_MONO tools=no target=release_debug
+  $SCONS platform=osx $OPTIONS $OPTIONS_MONO tools=no target=release
+
+  mkdir -p /root/out/templates-mono
+  cp -rvp bin/* /root/out/templates-mono
+  rm -rf bin
+
+  find /root/out -name config -exec cp /root/dependencies/mono/etc/config {} \;
+fi
+
+echo "macOS build successful"
